vendor_mgmt/requirements.txt
Django==4.2
djangorestframework==3.12.2
